+++ frontend/src/widgets/firmware-upgrade/FirmwareDevicesPanel.vue
@@ -12,6 +12,7 @@ import {
   saveDevice,
 } from './api'
 import type { Board, Device, FirmwareProfile } from './types'
+import ExternalFirmwarePanel from './ExternalFirmwarePanel.vue'
 
 defineEmits<{ close: [] }>()
 
@@ -354,6 +355,9 @@ onUnmounted(() => {
         </div>
         <p v-if="backupMsg" class="font-mono text-[10px] opacity-70">{{ backupMsg }}</p>
       </div>
+
+      <!-- External firmware: register + flash pre-built binaries -->
+      <ExternalFirmwarePanel />
     </template>
   </div>
 </template>